backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
python-dotenv==1.0.0
dj-database-url==2.1.0
requests==2.31.0
